refactor: remove ark-sponge dep from paramgen, permutation crates
This also adds a new crate `poseidon-consistency` in which we put
all the benchmarks and consistency property-tests for checking
ark-sponge vs. sage vs. this project consistency
