+++ README.md
@@ -5,6 +5,7 @@ This repository contains:
 * [`poseidon377`](../main/poseidon377): an instantiation of the Poseidon hash function for [`decaf377`](https://github.com/penumbra-zone/decaf377)
 * [`poseidon-paramgen`](../main/poseidon-paramgen): an independent implementation of Poseidon parameter generation
 * [`poseidon-permutation`](../main/poseidon-permutation): an independent implementation of the Poseidon permutation
+* [`poseidon-consistency`](../main/poseidon-consistency): property-based tests for consistency between Poseidon implementations
 
 ## Audits
