+++ poseidon-consistency/Cargo.toml
@@ -1,24 +1,29 @@
 [package]
-name = "poseidon-permutation"
+name = "poseidon-consistency"
 version = "0.1.0"
 edition = "2018"
 authors = ["Penumbra <crates@penumbra.zone>", "redshiftzero <jen@penumbra.zone>"]
-description = "An instantiation of the Poseidon permutation"
+description = "Checks consistency between Penumbra and Ark-sponge implementation"
 license = "MIT OR Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-once_cell = "1.8"
-ark-ff = "0.3"
 poseidon-paramgen = { version="0.1", path = "../poseidon-paramgen/" }
+poseidon-permutation = {version= "0.1", path="../poseidon-permutation/"}
+ark-sponge = { version="0.3", git = "https://github.com/penumbra-zone/sponge", branch = "r1cs" }
+decaf377 = "0.1"
 
 [dev-dependencies]
+criterion = { version = "0.3", features=["html_reports"] }
+ark-ff = "0.3"
 ark-ed-on-bls12-377 = "0.3"
 num-bigint = "0.4"
-ark-sponge = { version="0.3", git = "https://github.com/penumbra-zone/sponge", branch = "r1cs" }
-criterion = { version = "0.3", features=["html_reports"] }
-once_cell = "1.8"
 proptest = "1"
 rand_core = { version = "0.6.3", features = ["getrandom"] }
 rand_chacha = "0.3"
+once_cell = "1.8"
+
+[[bench]]
+name = "permutation"
+harness = false
