+++ poseidon-paramgen/Cargo.toml
@@ -10,7 +10,6 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 anyhow = "1"
-ark-sponge = { version="0.3", git = "https://github.com/penumbra-zone/sponge", branch = "r1cs" }
 ark-ff = "0.3"
 merlin = "3.0"
 num = "0.4"
