+++ poseidon-paramgen/src/lib.rs
@@ -38,7 +38,6 @@ pub use rounds::RoundNumbers;
 pub use utils::log2;
 
 use ark_ff::PrimeField;
-use ark_sponge::poseidon::PoseidonParameters as ArkPoseidonParameters;
 
 /// A set of Poseidon parameters for a given set of input parameters.
 #[derive(Clone, Debug)]
@@ -94,30 +93,6 @@ impl<F: PrimeField> PoseidonParameters<F> {
             arc,
             optimized_mds,
             optimized_arc,
-        }
-    }
-}
-
-impl<F: PrimeField> Into<ArkPoseidonParameters<F>> for PoseidonParameters<F> {
-    fn into(self) -> ArkPoseidonParameters<F> {
-        let alpha = match self.alpha {
-            Alpha::Exponent(exp) => exp as u64,
-            Alpha::Inverse => panic!("ark-sponge does not allow inverse alpha"),
-        };
-        // TODO: let user specify different capacity choices
-        let capacity = 1;
-        let rate = self.t - capacity;
-        let full_rounds = self.rounds.full();
-        let partial_rounds = self.rounds.partial();
-
-        ArkPoseidonParameters {
-            full_rounds,
-            partial_rounds,
-            alpha,
-            ark: self.arc.into(),
-            mds: self.mds.into(),
-            rate,
-            capacity,
         }
     }
 }
