+++ poseidon377/Cargo.toml
@@ -14,7 +14,7 @@ ark-ff = "0.3"
 poseidon-paramgen = { version="0.1", path= "../poseidon-paramgen" }
 poseidon-permutation = { version="0.1", path= "../poseidon-permutation" }
 ark-relations = { version="0.3", optional=true }
-ark-sponge = { version="0.3", git = "https://github.com/penumbra-zone/sponge", branch = "r1cs" }
+ark-sponge = { git = "https://github.com/penumbra-zone/sponge", branch = "r1cs" }
 decaf377 = {  version="0.1", features = ["r1cs"] }
 num-bigint = "0.4.3"
 ark-groth16 = { version = "0.3", optional=true }
